server/app/requirements.txt
eyed3==0.9.7
firebase-admin==6.2.0
Flask==2.3.3
Flask-Cors==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
markupsafe
argon2-cffi
